derive running command durations in history builder
